refactoring a bit after a little mess with merges
