--- src/Horarium.InMemory/InMemoryRepositoryFactory.cs
@@ -1,13 +1,0 @@
-using System;
-using Horarium.Repository;
-
-namespace Horarium.InMemory
-{
-    public static class InMemoryRepositoryFactory
-    {
-        private static readonly Lazy<InMemoryRepository> 
-            Repository = new Lazy<InMemoryRepository>(() => new InMemoryRepository());
-
-        public static IJobRepository Create() => Repository.Value;
-    }
-}
