+++ src/Horarium.InMemory/Indexes/FailedJobIndex.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using Horarium.Repository;
 
-namespace Horarium.InMemory.PerformantInMemory.Indexes
+namespace Horarium.InMemory.Indexes
 {
     public class FailedJobIndex : IAddRemoveIndex
     {
