+++ src/Horarium.InMemory/Indexes/ReadyJobIndex.cs
@@ -1,24 +1,28 @@
 using System;
 using System.Collections.Generic;
-using Horarium.InMemory.PerformantInMemory.Indexes.Comparers;
+using Horarium.InMemory.Indexes.Comparers;
 using Horarium.Repository;
 
-namespace Horarium.InMemory.PerformantInMemory.Indexes
+namespace Horarium.InMemory.Indexes
 {
-    public class RepeatJobIndex : IAddRemoveIndex
+    public class ReadyJobIndex : IAddRemoveIndex
     {
         private readonly SortedSet<JobDb> _startAtIndex = new SortedSet<JobDb>(new StartAtComparer());
         
+        private readonly JobKeyIndex _jobKeyIndex = new JobKeyIndex();
+
         public void Add(JobDb job)
         {
-            if (job.Status != JobStatus.RepeatJob) return;
+            if (job.Status != JobStatus.Ready) return;
             
             _startAtIndex.Add(job);
+            _jobKeyIndex.Add(job);
         }
 
         public void Remove(JobDb job)
         {
             _startAtIndex.Remove(job);
+            _jobKeyIndex.Remove(job);
         }
 
         public int Count()
@@ -32,6 +36,11 @@ namespace Horarium.InMemory.PerformantInMemory.Indexes
                 return _startAtIndex.Min;
 
             return null;
+        }
+
+        public JobDb GetJobKeyEqual(string jobKey)
+        {
+            return _jobKeyIndex.Get(jobKey);
         }
     }
 }
