+++ src/Horarium.InMemory/Indexes/RepeatJobIndex.cs
@@ -1,28 +1,24 @@
 using System;
 using System.Collections.Generic;
-using Horarium.InMemory.PerformantInMemory.Indexes.Comparers;
+using Horarium.InMemory.Indexes.Comparers;
 using Horarium.Repository;
 
-namespace Horarium.InMemory.PerformantInMemory.Indexes
+namespace Horarium.InMemory.Indexes
 {
-    public class ReadyJobIndex : IAddRemoveIndex
+    public class RepeatJobIndex : IAddRemoveIndex
     {
         private readonly SortedSet<JobDb> _startAtIndex = new SortedSet<JobDb>(new StartAtComparer());
         
-        private readonly JobKeyIndex _jobKeyIndex = new JobKeyIndex();
-
         public void Add(JobDb job)
         {
-            if (job.Status != JobStatus.Ready) return;
+            if (job.Status != JobStatus.RepeatJob) return;
             
             _startAtIndex.Add(job);
-            _jobKeyIndex.Add(job);
         }
 
         public void Remove(JobDb job)
         {
             _startAtIndex.Remove(job);
-            _jobKeyIndex.Remove(job);
         }
 
         public int Count()
@@ -36,11 +32,6 @@ namespace Horarium.InMemory.PerformantInMemory.Indexes
                 return _startAtIndex.Min;
 
             return null;
-        }
-
-        public JobDb GetJobKeyEqual(string jobKey)
-        {
-            return _jobKeyIndex.Get(jobKey);
         }
     }
 }
