+++ src/Horarium.IntegrationTest/TestParallelsWorkTwoManagers.cs
@@ -3,7 +3,6 @@ using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
 using Horarium.InMemory;
-using Horarium.InMemory.PerformantInMemory;
 using Horarium.IntegrationTest.Jobs;
 using Horarium.Interfaces;
 using Horarium.Mongo;
@@ -19,7 +18,6 @@ namespace Horarium.IntegrationTest
         {
             MongoDB,
             InMemory,
-            PerformantInMemory
         }
 
         public TestParallelsWorkTwoManagers()
@@ -41,10 +39,6 @@ namespace Horarium.IntegrationTest
                     break;
                 
                 case DataBase.InMemory:
-                    jobRepository = InMemoryRepositoryFactory.Create();
-                    break;
-                
-                case DataBase.PerformantInMemory:
                     jobRepository = new PerformantInMemoryRepository();
                     break;
                 
@@ -58,34 +52,9 @@ namespace Horarium.IntegrationTest
             return horarium;
         }
 
-        [Fact]
-        public async Task Test123()
-        {
-            var repository = new PerformantInMemoryRepository();
-            var client = new HorariumClient(repository);
-
-            await client.Create<TestJob, TestJobParam>(new TestJobParam
-            {
-                Counter = 123,
-                DbType = DataBase.PerformantInMemory
-            }).Schedule();
-
-            try
-            {
-                var job = await repository.GetReadyJob("we", TimeSpan.FromSeconds(10));
-            }
-            catch (Exception e)
-            {
-                var c = 4;
-            }
-
-            var a = 5;
-        }
-
         [Theory]
-        //[InlineData(DataBase.MongoDB)]
-        //[InlineData(DataBase.InMemory)]
-        [InlineData(DataBase.PerformantInMemory)]
+        [InlineData(DataBase.MongoDB)]
+        [InlineData(DataBase.InMemory)]
         public async Task TestParallels(DataBase dataBase)
         {
             var firstScheduler = CreateScheduler(dataBase);
@@ -130,7 +99,7 @@ namespace Horarium.IntegrationTest
          * synchronization of shared static resources becomes VERY problematic
          */ 
         //[InlineData(DataBase.InMemory)]
-        [InlineData(DataBase.PerformantInMemory)]
+        [InlineData(DataBase.InMemory)]
         public async Task Scheduler_SecondInstanceStart_MustUpdateRecurrentJobCronParameters(DataBase dataBase)
         {
             var watch = Stopwatch.StartNew();
